+++ packages/eslint-config/react.js
@@ -6,7 +6,7 @@
  */
 
 const rules = [
-  './rules/deprecated/react.js',
+  './rules/deprecated/react',
   './rules/react/a11y',
   './rules/react/base',
   './rules/react/hooks',